requirements.txt
click>=7.1.1
cycler>=0.10.0
Flask>=1.1.2
itsdangerous>=1.1.0
jieba>=0.42.1
Jinja2>=2.11.1
joblib>=0.14.1
kiwisolver>=1.2.0
MarkupSafe>=1.1.1
matplotlib>=3.2.1
nltk>=3.4.5
numpy>=1.18.2
Pillow>=7.1.1
pyparsing>=2.4.7
pytesseract>=0.3.3
python-dateutil>=2.8.1
scikit-learn>=0.22.2.post1
scipy>=1.4.1
six>=1.14.0
Werkzeug>=1.0.1
wordcloud>=1.6.0
gensim>=3.8.2
